backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
structlog==23.2.0
pydantic==2.4.2
hyperliquid-python-sdk==0.22.0
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
